move logic to testing package
