+++ backend/src/services/testing-config/index.ts
@@ -2,20 +2,14 @@ import { MetloContext } from "types"
 import { UpdateTestingConfigParams } from "@common/api/testing-config"
 import { AppDataSource } from "data-source"
 import { createQB, getQB, insertValueBuilder } from "services/database/utils"
-import { ApiEndpoint, DataField, TestingConfig } from "models"
+import { ApiEndpoint, TestingConfig } from "models"
 import { TestingConfigResp } from "@common/types"
-import { parseResourceConfig, ResourceConfigParseRes } from "@metlo/testing"
+import {
+  parseResourceConfig,
+  ResourceConfigParseRes,
+  findEndpointResourcePermissions,
+} from "@metlo/testing"
 import { RedisClient } from "utils/redis"
-import { DataSection } from "@common/enums"
-
-const DATA_SECTION_TO_PATH: Record<DataSection, string> = {
-  [DataSection.REQUEST_PATH]: "req.path",
-  [DataSection.REQUEST_QUERY]: "req.query",
-  [DataSection.REQUEST_HEADER]: "req.header",
-  [DataSection.REQUEST_BODY]: "req.body",
-  [DataSection.RESPONSE_HEADER]: "res.header",
-  [DataSection.RESPONSE_BODY]: "res.body",
-}
 
 export const getTestingConfig = async (
   ctx: MetloContext,
@@ -126,94 +120,25 @@ export const getEntityTagsCached = async (
 const getResourcePermissions = async (
   ctx: MetloContext,
   endpoint: ApiEndpoint,
-  dataFields: DataField[],
 ): Promise<string[]> => {
   const config = await getTestingConfigCached(ctx)
-  if (!config?.configString) {
+  try {
+    return findEndpointResourcePermissions(endpoint, config?.configString)
+  } catch {
     return []
   }
-  const parseRes = parseResourceConfig(config.configString)
-  if (!parseRes.res) {
-    throw new Error("Bad Testing Config")
-  }
-  const configObject = parseRes.res
-  const resoucePermissions = new Set<string>()
-  const endpointEntities: { type: string; path: string }[] = []
-  for (const dataField of dataFields) {
-    if (dataField.entity) {
-      endpointEntities.push({
-        type: dataField.entity.split(".")?.[0],
-        path: DATA_SECTION_TO_PATH[dataField.dataSection],
-      })
-    }
-  }
-
-  for (const item of configObject) {
-    if (item.type == "resource") {
-      const permFilters = item.members?.endpoints
-      if (permFilters && permFilters.length > 0) {
-        for (const permFilter of permFilters) {
-          if (permFilter.contains_resource) {
-            const containsResource = permFilter.contains_resource
-            if (
-              endpointEntities.some(
-                e =>
-                  ((containsResource.type &&
-                    e.type === containsResource.type) ||
-                    (!containsResource.type && e.type === item.name)) &&
-                  e.path.startsWith(containsResource.path),
-              )
-            ) {
-              permFilter.permissions.forEach(perm => {
-                resoucePermissions.add(`${item.name}.${perm}`)
-              })
-            }
-          } else {
-            if (permFilter.method) {
-              if (typeof permFilter.method === "string") {
-                if (permFilter.method !== endpoint.method) {
-                  continue
-                }
-              } else {
-                if (!permFilter.method.includes(endpoint.method)) {
-                  continue
-                }
-              }
-            }
-            if (
-              permFilter.host &&
-              !endpoint.host.match(new RegExp(permFilter.host))
-            ) {
-              continue
-            }
-            if (
-              permFilter.path &&
-              !endpoint.path.match(new RegExp(permFilter.path))
-            ) {
-              continue
-            }
-            permFilter.permissions.forEach(perm => {
-              resoucePermissions.add(`${item.name}.${perm}`)
-            })
-          }
-        }
-      }
-    }
-  }
-  return [...resoucePermissions]
 }
 
 export const getResourcePermissionsCached = async (
   ctx: MetloContext,
   endpoint: ApiEndpoint,
-  dataFields: DataField[],
 ): Promise<string[]> => {
   const key = `endpointResourcePermissionsCached:e#${endpoint.uuid}`
   const cacheRes: string[] | null = await RedisClient.getFromRedis(ctx, key)
   if (cacheRes !== null) {
     return cacheRes
   }
-  const realRes = await getResourcePermissions(ctx, endpoint, dataFields)
+  const realRes = await getResourcePermissions(ctx, endpoint)
   await RedisClient.addToRedis(ctx, key, realRes, 60)
   return realRes
 }
