+++ testing/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@metlo/testing",
-  "version": "0.3.6",
+  "version": "0.3.7",
   "license": "MIT",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
