+++ testing/src/index.ts
@@ -145,3 +145,4 @@ export { TestTemplate } from "./templates/types"
 export { TemplateConfig, ResourceConfigParseRes } from "./types/resource_config"
 export { runTest, estimateTest } from "./runner"
 export { generateAuthTests } from "./generate/generate-auth-tests"
+export { findEndpointResourcePermissions } from "./generate/permissions"
